commit 3

+++ resources/views/users/index.blade.php
@@ -3,17 +3,19 @@
 <div class="row">
     <div class="col-lg-12 margin-tb">
         <div class="pull-left">
-            <h2>Users Management</h2>
+            <h2>User Management</h2>
         </div>
         <div class="pull-right">
-            <a class="btn btn-success" href="{{ route('users.create') }}"> Create New User</a>
+        @can('user-create')
+            <a class="btn btn-success" href="{{ route('users.create') }}"> Create New user</a>
+            @endcan
         </div>
     </div>
 </div>
 @if ($message = Session::get('success'))
-<div class="alert alert-success">
-    <p>{{ $message }}</p>
-</div>
+    <div class="alert alert-success">
+        <p>{{ $message }}</p>
+    </div>
 @endif
 <table class="table table-bordered">
     <tr>
@@ -23,7 +25,8 @@
         <th>Roles</th>
         <th width="280px">Action</th>
     </tr>
-@foreach ($data as $key => $user)
+    
+    @foreach ($data as $key => $user)
     <tr>
         <td>{{ ++$i }}</td>
         <td>{{ $user->name }}</td>
@@ -34,16 +37,20 @@
                     <span class="badge rounded-pill bg-dark">{{ $v }}</span>
                 @endforeach
             @endif
-        </td>
-        <td>
+            </td>
+            <td>
             <a class="btn btn-info" href="{{ route('users.show',$user->id) }}">Show</a>
-            <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+            @can('user-edit')
+                <a class="btn btn-primary" href="{{ route('users.edit',$user->id) }}">Edit</a>
+            @endcan
+            @can('user-delete')
                 {!! Form::open(['method' => 'DELETE','route' => ['users.destroy', $user->id],'style'=>'display:inline']) !!}
                     {!! Form::submit('Delete', ['class' => 'btn btn-danger']) !!}
                 {!! Form::close() !!}
+            @endcan
         </td>
     </tr>
-@endforeach
+    @endforeach
 </table>
 {!! $data->render() !!}
 @endsection
